Working with limelight
(ian) I know have a half-decent idea of what I am doing with this.  We need a method in the robot class and a couple other things, but I can have Owen and Aaron get those done, as it is relatively simple to do. I have also found some useful examples on the documentation website so actions such as "chasing" a target and etc can be done.